Basic Mailing Impl (untested) & couple bug fixes

+++ src/ghHook.go
@@ -5,24 +5,26 @@
 package main
 
 import (
+	"bytes"
 	"encoding/json"
 	"flag"
 	"fmt"
 	"io"
 	"log"
 	"net/http"
+	"net/smtp"
 	"os"
 	"os/exec"
 	"strings"
+	"text/template"
 )
 
+// Global config Variable, gets initialized in main()
 var config Config
 
 // Used to access GitHub JSON Payload from the program
 // go JSON Unmarshal ignores case of members
 type ghPayload struct {
-	// TODO get from JSON body
-	HeaderDeliveryId string
 	// which branch was pushed to
 	Ref string
 	// commit id before push
@@ -40,19 +42,76 @@ type ghPayload struct {
 	CompareURL string `json:"compare"`
 }
 
-func RunRoutine(payload *ghPayload) {
+// Send Test Execution Report to recipients
+// ${pusher} inside of recipients will be replaced with payload.Pusher.email
+func SendReportMail(payload *ghPayload, recipients []string, success bool) {
+
+	if config.DoDebug {
+		fmt.Printf("Report: payload: %+v\n", *payload)
+	}
+
+	// Fill the email body template
+	templ, templErr := template.ParseFiles("emailTempl.html")
+
+	if templErr != nil {
+		log.Printf("Error parsing html email template <%v>, wont send mail", templErr)
+		return
+	}
+
+	subject := "Test %s on " + payload.Ref
+	if success {
+		subject = fmt.Sprintf(subject, "successfull")
+	} else {
+		subject = fmt.Sprintf(subject, "FAILED")
+	}
+
+	var mailBody bytes.Buffer
+
+	mimeHeaders := "MIME-version: 1.0;\nContent-Type: text/html; charset=\"UTF-8\";\n\n"
+	mailBody.Write([]byte(fmt.Sprintf("Subject: %s \n%s\n\n", subject, mimeHeaders)))
+
+	if templFillErr := templ.Execute(&mailBody, *payload); templFillErr != nil {
+		log.Printf("Couldnt fill email body template <%v>", templFillErr)
+		log.Println("Defaulting back to %+v representation; Some data might have still been written and the result may appear scuffed")
+		mailBody.WriteString(fmt.Sprintf("%+v", *payload))
+	}
+
+	// replace placeholders in recipients
+	for idx, recp := range recipients {
+		if recp == "${pusher}" {
+			recipients[idx] = payload.Pusher.email
+		}
+	}
+
+	// Instatiate SMTP Auth interface
+	smtpAuth := smtp.PlainAuth("", config.Mailing.FromAddr, config.Mailing.FromAddr, config.Mailing.SMTPHost)
+
+	mailErr := smtp.SendMail(config.Mailing.SMTPHost+":"+config.Mailing.SMTPPort, smtpAuth, config.Mailing.FromAddr, recipients, []byte(fmt.Sprintf("%+v", payload)))
+
+	if mailErr != nil {
+		log.Printf("Error sending report mail <%v>", mailErr)
+	}
+
+}
+
+// Actual Logic of the Program
+// Clones the repo (only branch that was pushed to) living at payload.CloneURL into config.Routines[i].GitCloneDir
+// it does this for EVERY element in config.Routines where the branch is included in routine.BranchesToProcess
+func RunRoutine(payload *ghPayload, deliveryID string) {
 	// find the correct routine
 	splitRefPath := strings.Split(payload.Ref, "/")
-	branch := splitRefPath[len(splitRefPath)]
+	branch := splitRefPath[len(splitRefPath)-1] // last element should be only the branch name
 
 	for _, routine := range config.Routines {
 		for _, branchToDo := range routine.BranchesToProcess {
 			if branchToDo == branch {
+				// guaranteed to be unique as long as github works correctly (and the correct argument was passed :) )
+				gitCloneDir := routine.GitCloneDir + deliveryID + branch
 
-				gitCloneDir := routine.GitCloneDir + payload.HeaderDeliveryId
+				// TODO debug here
 
 				if cloneDirErr := os.Mkdir(gitCloneDir, os.ModePerm); cloneDirErr != nil {
-					log.Printf("Error creating clone directory(<%s>), wont process this payload <%s>\n", payload.HeaderDeliveryId, cloneDirErr)
+					log.Printf("Error creating clone directory(<%s>), wont process this payload <%v>\n", deliveryID, cloneDirErr)
 					return
 				}
 				// Use git to clone the branch that was pushed to
@@ -62,7 +121,7 @@ func RunRoutine(payload *ghPayload) {
 				// stderr gets merged to stdout, runs command
 				stdout, gitErr := gitCloneCmd.CombinedOutput()
 				if gitErr != nil {
-					log.Printf("Error cloning from git <%s>, skipping this one\n", gitErr)
+					log.Printf("Error cloning from git <%v>, skipping this one\n", gitErr)
 					return
 				}
 
@@ -71,23 +130,34 @@ func RunRoutine(payload *ghPayload) {
 					log.Println(stdout)
 				}
 
+				// Run tests on the cloned repo
 				testCmd := exec.Command(routine.TestCloneCmdExe, routine.TestCloneCmdArgs)
 				testCmd.Dir = gitCloneDir
 				testoutput, testErr := testCmd.CombinedOutput()
 				if testErr != nil {
 					log.Printf("ERROR running Test on branch <%s> by %s (%s)\n", branch, payload.Pusher.name, payload.Pusher.email)
-					log.Printf("ERROR description <%s>\n", testErr)
+					log.Printf("ERROR description <%v>\n", testErr)
 					log.Println("Command output:")
 					log.Println(testoutput)
-					// TODO SendMail(...)
+
+					if routine.RemoveOnFailure {
+						// remove clone after successfull test
+						defer os.RemoveAll(gitCloneDir)
+					}
+					SendReportMail(payload, routine.MailTo, false)
 				} else {
 					log.Printf("Test Successfull on branch <%s> by %s (%s)\n", branch, payload.Pusher.name, payload.Pusher.email)
+
 					if config.DoDebug {
 						log.Println("Command output:")
 						log.Println(testoutput)
 					}
-					// remove clone after successfull test
-					defer os.RemoveAll(gitCloneDir)
+
+					if routine.RemoveOnSuccess {
+						// remove clone after successfull test
+						defer os.RemoveAll(gitCloneDir)
+					}
+					SendReportMail(payload, routine.MailTo, true)
 				}
 
 			}
@@ -101,7 +171,7 @@ func HookHandler(httpResp http.ResponseWriter, httpReq *http.Request) {
 	var HandleFail = func(reqStatus int, errMsg string) {
 		httpResp.WriteHeader(reqStatus)
 		io.WriteString(httpResp, "{}")
-		log.Fatalln(errMsg)
+		log.Fatalln(errMsg) // Fatal will quit the program, following returns are simply for readability
 	}
 
 	if config.DoDebug {
@@ -112,32 +182,34 @@ func HookHandler(httpResp http.ResponseWriter, httpReq *http.Request) {
 		}
 	}
 
-	hookCtx, err := ParseHook([]byte(config.GitHubHookSecret), httpReq)
-
 	httpResp.Header().Set("Content-type", "application/json")
 
-	if err != nil {
-		HandleFail(http.StatusBadRequest, fmt.Sprintf("ERROR processing hook <%s>\n", err))
+	hookCtx, hookErr := ParseHook([]byte(config.GitHubHookSecret), httpReq)
+
+	if hookErr != nil {
+		HandleFail(http.StatusBadRequest, fmt.Sprintf("ERROR processing hook <%v>\n", hookErr))
 		return
 	}
 
-	log.Printf("Received event <%s>", hookCtx.Event)
+	log.Printf("Received event <%s>\n", hookCtx.Event)
 
 	if config.DoDebug {
 		log.Printf("Payload <\n%s\n>\n", hookCtx.Payload)
 	}
 
-	// parse `hookCtx.Payload` or do additional processing here
+	// payload parsing
 	if hookCtx.Event != "push" {
 		log.Printf("Wont process event type<%s>", hookCtx.Event)
 		httpResp.WriteHeader(http.StatusNotImplemented)
 	} else {
 		var recvPayload ghPayload
 		if jsonErr := json.Unmarshal(hookCtx.Payload, &recvPayload); jsonErr == nil {
-			RunRoutine(&recvPayload)
+			// Actual programm Logic
+			RunRoutine(&recvPayload, hookCtx.Id)
+
 			httpResp.WriteHeader(http.StatusOK)
 		} else { // Error in JSON Request body from GitHub
-			HandleFail(http.StatusBadRequest, fmt.Sprintf("ERROR unmarshalling JSON request Payload <%s>", err))
+			HandleFail(http.StatusBadRequest, fmt.Sprintf("ERROR unmarshalling JSON request Payload <%v>", jsonErr))
 			return
 		}
 	}
@@ -148,14 +220,18 @@ func HookHandler(httpResp http.ResponseWriter, httpReq *http.Request) {
 // Program entry point
 func main() {
 	//by default look for the config file in the directory the program was executed, if its a symlink it might be invalid by now
-	exe, exeErr := os.Executable()
+	exe, exeErr := os.Getwd()
 	if exeErr != nil {
-		panic("Unable to determine executable directory")
+		panic(fmt.Sprintf("Unable to determine executable directory <%v>", exeErr))
 	}
+	exe += "/conf.json"
+	log.Printf("Picking up default config at %s\n", exe)
 	configArg := flag.String("config", exe, "path to the JSON config file")
-	flag.Parse()
-	config = LoadConfig(*configArg)
+	flag.Parse() // automatically handles -help, unhandled flags, etc...
 
+	config = LoadConfig(*configArg)
+	// spawns a new go routine for each request
+	// access to 'config' still IS thread safe as it wont be modified after loading
 	http.HandleFunc("/payload", HookHandler)
 
 	http.ListenAndServe(":8081", nil)
